+++ src/attack/attack.ts
@@ -1,9 +1,10 @@
 import { Character } from "../character/Character";
-import { d20, adjustHP } from "../gameState";
+import { d20 } from "../gameState";
 import { getCharacter } from "../character/getCharacter";
 import { AttackResult } from "./AttackResult";
 import { getCharacterStatModified } from "../character/getCharacterStatModified";
 import { updateCharacterQuestProgess } from "../quest/updateQuestProgess";
+import { adjustHP } from "../character/adjustHP";
 
 export const attack = (
   attackerId: string,
