+++ src/character/equipItem.ts
