+++ src/commands/encounters/chest.ts
@@ -4,7 +4,9 @@ import {
   MessageAttachment,
   MessageEmbed,
 } from "discord.js";
-import { adjustHP, adjustGold, awardXP } from "../../gameState";
+import { adjustGold } from "../../character/adjustGold";
+import { adjustHP } from "../../character/adjustHP";
+import { awardXP } from "../../character/awardXP";
 import { getUserCharacter } from "../../getUserCharacter";
 import { updateStatusEffect } from "../../statusEffects/grantStatusEffect";
 import { trapAttack } from "../../trap/trap";
